// src/modules/inventory/reports/yarn-store/yarn-delivery-challan-gate-pass-report/yarn-delivery-challan-gate-pass-report-type.ts
export interface IYarnDeliveryChallanGatePassReport {
  ID: number;
  COMPANY_ID: number;
  COMPANY_NAME: string;
  PREFIX: string;
  ADDRESS: string;
  CONTACT: string;
  EMAIL: string;
  WEB: string;
  REMARKS: string;
  IMAGE: string;
  ADVICE: string;
  TREATMENT: string;
  BUYER_ID: number;
  BUYER: string;
  PO_ID: number;
  PO_NO: string;
  CHALLAN_NO: string;
  CHALLAN_DATE: Date;
  ISSUE_TYPE_ID: number;
  ISSUE_TYPE: string;
  KNITTING_HOUSE_ID: number;
  KNITTING_HOUSE: string;
  KNITTING_HOUSE_ADDRESS: string;
  MODEL_NO_ID: number;
  STYLE: string;
  STYLE_ID: string;
  GSM: string;
  FABRIC_TYPE_ID: number;
  FABRIC_TYPE: string;
  FABRIC_ID: number;
  FABRIC: string;
  BB_LC_NO: string;
  WORK_ORDER_NUMBER: string;
  YARN_BRAND: string;
  YARN_LOT_NO_ID: number;
  YARN_LOT_NUMBER: string;
  YARN_ID: number;
  YARN: string;
  YARN_COLOR_ID: number;
  YARN_COLOR: string;
  FABRIC_PART_ID: number;
  FABRIC_PART: string;
  GMT_COLOR_ID: number;
  GMT_COLOR: string;
  GMT_SIZE_ID: number;
  GMT_SIZE: string;
  DIA: string;
  SHAPE: string;
  STITCH_LENGTH: string;
  PER_CTN_QTY: number;
  COLLAR_CUFF_RATIO: string;
  CARTON_QTY: number;
  PICES: number;
  QUANTITY: number;
  GAUGE: string;
  OPEN_DIA: string;
  COMPOSED_BY: string;
  SORTING_NO: string;
  CONE: number;
  KNITTING_PROGRAM_NO: string;
  MAIN_COMPANY: string;
  MAIN_COMPANY_ADDRESS: string;
}